Module uic_upti_writer::log_formatter

Structs

DebugData

Constants

AEM_COUNTER

AEM_CURRENT_PACKET

AEM_CURRENT_PACKET_V2

AEM_REQUEST

AEM_RESPONSE

AEM_SAMPLE

API_RX

API_TRACE

API_TX

ASSERT

CATEGORY_BACKCHANNEL

CATEGORY_UNKNOWN

CORE_DUMP

CPU_USAGE

DAG_TRACE

EEPROM_READ_REQUEST

EEPROM_READ_RESPONSE

EEPROM_WRITE_REQUEST

EEPROM_WRITE_RESPONSE

ERROR

EXCEPTION_PACKET

EZSP

EZSP_UART

FLASH_READ_REQUEST

FLASH_READ_RESPONSE

HEAP_STACK

INFO_REQUEST

INFO_RESPONSE

INFO_ZIP_VERSION

INFO_ZNET_VERSION

LATENCY

MASK_AEM

MASK_PACKET

MUSTANG

PACKET_TRACE

PACKET_TRACE_EFR_OTHER

PACKET_TRACE_EFR_RX

PACKET_TRACE_EFR_TX

PACKET_TRACE_EM2XX_RX

PACKET_TRACE_EM2XX_TX

PACKET_TRACE_EM3XX_RX

PACKET_TRACE_EM3XX_TX

PACKET_TRACE_EM2420_RX

PACKET_TRACE_EM2420_TX

PACKET_TRACE_PRO2P_OTHER

PACKET_TRACE_PRO2P_RX

PACKET_TRACE_PRO2P_TX

PC_SAMPLE_PACKET

PHY_RX

PHY_TX

PRINTF

RADIO_REBOOT_COUNT

RAM_READ_REQUEST

RAM_READ_RESPONSE

RAM_WRITE_REQUEST

RAM_WRITE_RESPONSE

RESET_INFO

SIMULATED_EZSP_CALLBACK_READY

SIMULATED_NCP_IS_AWAKE

SIMULATED_WAKEUP_NCP

STATS

TIME

TIME_SYNC

TIME_SYNC_TEST

TMSP

USER_COMMAND

USER_RESPONSE

VIRTUAL_UART_RX

VIRTUAL_UART_TX

Statics

CLOSE_BRACKET

OPEN_BRACKET

Functions

get_full_debug_type

Returns the event name and event id based on the event type. Match is generated from debug-message-type.json, feature date “2022.1.8”

log_data_from_raw

Constructs formatted log record from the raw data

parse_raw

Constructs a DebugData object from the raw message bytes. Assumes that the message has been processed by the backchannel board (so that it is in the ethernet format rather than the serial format) and that the framing and length bytes have been removed.