Struct uic_upti_writer::mqtt_handler::__mock_MockMqttHandlerProcessTrait_MqttHandlerProcessTrait::__process_adapters_list::Common

struct Common {
    matcher: Mutex<Matcher>,
    seq_handle: Option<SeqHandle>,
    times: Times,
}

Holds the stuff that is independent of the output type

Fields
matcher: Mutex<Matcher>seq_handle: Option<SeqHandle>times: Times
Implementations

impl Common

fn call(&self, desc: &str)

fn in_sequence(&mut self, __mockall_seq: &mut Sequence) -> &mut Self

fn is_done(&self) -> bool

fn matches(&self, json: &JsonValue) -> bool

fn never(&mut self)

Forbid this expectation from ever being called.

fn satisfy_sequence(&self)

fn times<MockallR>(&mut self, __mockall_r: MockallR) where
MockallR: Into<TimesRange>,

Expect this expectation to be called any number of times contained with the given range.

fn with<MockallMatcher0: Predicate<JsonValue> + Send + 'static>(
&mut self,
json: MockallMatcher0
)

fn withf<MockallF>(&mut self, __mockall_f: MockallF) where
MockallF: Fn(&JsonValue) -> bool + Send + 'static,

fn withf_st<MockallF>(&mut self, __mockall_f: MockallF) where
MockallF: Fn(&JsonValue) -> bool + 'static,

fn verify_sequence(&self, desc: &str)